reorganize helpers and builders folders (#438)
