+++ SkeletonViewCore/Sources/API/Models/SkeletonTransitionStyle.swift
