+++ SkeletonViewCore/Sources/Internal/SkeletonLayerBuilders/SkeletonMultilineLayerBuilder.swift
@@ -5,6 +5,7 @@ import UIKit
 /// Object that facilitates the creation of skeleton layers for multiline
 /// elements, based on the builder pattern
 class SkeletonMultilineLayerBuilder {
+    
     var skeletonType: SkeletonType?
     var index: Int?
     var height: CGFloat?
@@ -93,4 +94,5 @@ class SkeletonMultilineLayerBuilder {
 
         return layer
     }
+    
 }
