+++ SkeletonViewCore/Sources/Internal/UIKitExtensions/UIView+Transitions.swift
@@ -2,25 +2,13 @@
 
 import UIKit
 
-extension CALayer {
-    func insertSkeletonLayer(_ sublayer: SkeletonLayer, atIndex index: UInt32, transition: SkeletonTransitionStyle, completion: (() -> Void)? = nil) {
-        insertSublayer(sublayer.contentLayer, at: index)
-        switch transition {
-        case .none:
-            completion?()
-        case .crossDissolve(let duration):
-            sublayer.contentLayer.setOpacity(from: 0, to: 1, duration: duration, completion: completion)
-        }
-    }
-}
-
 extension UIView {
+    
     func startTransition(transitionBlock: @escaping () -> Void) {
         guard let transitionStyle = _currentSkeletonConfig?.transition,
-            transitionStyle != .none
-            else {
-                transitionBlock()
-                return
+              transitionStyle != .none else {
+            transitionBlock()
+            return
         }
         
         if case let .crossDissolve(duration) = transitionStyle {
@@ -31,4 +19,5 @@ extension UIView {
                               completion: nil)
         }
     }
+    
 }
